base header tiny changes

+++ include/foundation/tools/time.hpp
@@ -8,7 +8,7 @@
 
 namespace AMTime {
 double seconds();
-int64_t miliseconds();
+double miliseconds();
 std::string Str(std::string_view format = "%Y-%m-%d %H:%M:%S",
                 std::optional<double> timestamp = std::nullopt);
 std::chrono::steady_clock::time_point SteadyNow();
